# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Animal Crossing data galore",
        "title": "Nookbase",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/search/{category}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Performs a search for items in the given category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Item category to search through (eg: fish, insects, etc.)",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/categories.CategoryItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "categories.CategoryItem": {
            "type": "object",
            "properties": {
                "catch_difficulty": {
                    "type": "string"
                },
                "catchphrase": {
                    "type": "string"
                },
                "critterpedia_image": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "furniture_image": {
                    "type": "string"
                },
                "icon_image": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "movement_speed": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nh_availability": {
                    "$ref": "#/definitions/categories.YearlyAvailability"
                },
                "number": {
                    "type": "integer"
                },
                "sell": {
                    "type": "integer"
                },
                "sh_availability": {
                    "$ref": "#/definitions/categories.YearlyAvailability"
                },
                "shadow": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "spawn_rates": {
                    "type": "string"
                },
                "surface": {
                    "type": "string"
                },
                "total_catches_to_unlock": {
                    "type": "integer"
                },
                "unique_id": {
                    "type": "string"
                },
                "vision": {
                    "type": "string"
                },
                "weather": {
                    "type": "string"
                },
                "where_how": {
                    "type": "string"
                }
            }
        },
        "categories.YearlyAvailability": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid input"
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        }
    }
}